+++ cmake/modules/Element.cmake
@@ -40,6 +40,7 @@ target_sources(kwik_element
             modules/element/lazy_list.cppm
             modules/element/keyboard.cppm
             modules/element/datepicker.cppm
+            modules/element/chart.cppm
 
         BASE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/modules/layout
         FILES
@@ -87,6 +88,7 @@ target_sources(kwik_element
         src/element/lazy_list.cpp
         src/element/keyboard.cpp
         src/element/datepicker.cpp
+        src/element/chart.cpp
 
         src/layout/flex_layout.cpp
         src/layout/grid_layout.cpp